modfied person endpoint combined credit respone to include fields that are missing on the tmdb api website but exist inthe data the api returns. Added fields that were found in person tv credits to the person combined credits as optional fields, and made the name field that previously existed optional as well
