+++ packages/core/src/types/v3/people/get-details.ts
@@ -1,7 +1,6 @@
 import { AppendToResponseType, PeopleAppendToResponse } from "../../append-to-response";
 
 export interface PeopleGetDetailsBaseResponse {
-
   birthday: string;
   known_for_department: string;
   deathday: string | null;
