+++ packages/core/src/types/v3/people/get-tv-credits.ts
@@ -1,17 +1,9 @@
-import { PeopleBaseCastGetTvCredit, PeopleBaseCrewGetTvCredit } from "./base-cast-crew-credits";
+import { PeopleBaseCastGetTvCredit, PeopleBaseCrewGetTvCredit, PeopleGetTvCreditsPerson } from "./base-cast-crew-credits";
 
 export interface PeopleGetTvCreditsResponse {
   cast: PeopleGetTvCreditsCast[];
   crew: PeopleGetTvCreditsCrew[];
   id: number;
-}
-
-export interface PeopleGetTvCreditsPerson {
-  origin_country: string[];
-  original_name: string;
-  first_air_date: string;
-  name: string;
-  episode_count: number;
 }
 
 export type PeopleGetTvCreditsCast = PeopleGetTvCreditsPerson & PeopleBaseCastGetTvCredit;
